Make TypeAlias.ty private
Summary:
Makes the type held by a TypeAlias private and adds functions `as_value()` and `as_type()` to access the type. The next diff will add a check inside `as_value()` for using a scoped type alias as a value (which is illegal).

This is less error-prone than allowing callers to access the type directly and do their own error checks, which they may forget to do.

Reviewed By: stroxler

Differential Revision: D66265807

fbshipit-source-id: 2cbf77aa1e2cc536d2639882edb3b0a59f6ae9e2