Use richer type information on `ClassType` returning functions
Summary:
I ran into a scenario where I needed access to the `ClassType` itself
resulting from promotions during a refactor, and I realized that in general
it's probably better to return `ClassType` directly and let the caller
wrap it in a `Type` because this makes the type signatures more informative.

The caller code is now a bit more verbose, but I'll add a helper to clean
it back up in the next diff.

Reviewed By: grievejia

Differential Revision: D66337487

fbshipit-source-id: b571d3583da23dddfcfb409aa6ceb68d8d2b44e1