Make a `Type::type_form` helper to reduce boilerplate
Summary:
Creating a valid `Type::Type` is a bit noisy - you have to
add two levels of parentheses and a `Box::new(` call, all of which
is pure boilerplate.

This commit adds a helper to reduce the boilerplate, partly as
a penance for my previous diff having made `ClassType` creation a bit
more verbose.

Since we cannot name a function `type`, I chose `type_form`, because
`Type::Type` corresponds almost exactly to PEP 747 TypeForm, but I don't
feel strongly about this name, I just think it's good if we start moving
away from confusing `Type::Type` with the python `type[...]` construct,
which is actually a generic class (it's the base class of class definition
values, and is very much not the same thing as "any type expression").

Reviewed By: grievejia

Differential Revision: D66337488

fbshipit-source-id: 4eb7f7ac54e5f5734bdf97b5f45c3c25c99537cf