Discard type hint If contextual type is invalid to begin with
Summary:
Sometimes contextual hint can be invalid. For example,
```
x: int = [0, 1, 2]  # The hint `int` is not a super type of list
```
When we hit this case, what should we do? Previously, we still stay in contextual typing mode, using the type `Any` as the hint to type each individual elements. Although it's not an incorrect thing to do, I found the logic a little bit confusing. Here I'm adjusting the code a little bit so that when the hint is invalid, we don't do contextual typing on the elements anymore, and switch to full inference mode instead. It shouldn't meaningfully change the semantics of the type system -- all it's going to affect is some error messages (in a good way, in the sense that the "actual type" of the expressions now becomes more clear than just the hint type `Any`).

Reviewed By: samwgoldman

Differential Revision: D66339955

fbshipit-source-id: e9b19ca60b354d32680fcfc37e916e6fb795d4da