Move self type binding to near the end.
Summary:
It is risky to let anything related to basic class analysis depend
on `self_type`, and it's usually not necessary - we can depend on
the class definition instead.

(The problem comes from the fact that we have to analyze type parameters
to compute `self_type`, which can create recursion issues).

Moving the `self_type` code, which used to be near the top because things
like base classes and mro used to rely on it, to the bottom to make it
clearer that no bindings actually need it.

Reviewed By: ndmitchell

Differential Revision: D66457375

fbshipit-source-id: eec5b68be7e555ab00e549f05c19e1a1e0a99cc3