Allow invariant violations on `ClassType`
Summary:
Legacy type parameters present a significant architectural challenge for
Pyre2:
- for sanity of the overall type checker, we very much want `ClassType` to
  always have valid type arguments. This means indexing a class name requires
	looking up the type parameters.
- type parameters cannot be determined without looking at base classes, which
  consist of a list of expressions. Without forking a large fraction of Pyre2's
	main expression switch to special-case for type expressions, our main option
	is to use `infer_expr` on the bases.
- but a base class expression can contain the current class; in 'bad' user code
  we could do all sorts of things like try to inherit from ourselves, but it can
	happen even in 'good' user code, for example `class Foo(list[Foo[T]])`. There's
	an example like this in `typeshed`, the `_ScandirIterator` type.

The problem that happens is we hit recursive `tparams` lookup, which currently
returns empty params. We'll complain that `Foo` takes no arguments, and
although the type params lookup itself isn't cached we cache a base class type
of `list[Foo]`.

Later, if we ever try to actually use that type in a way that requires a
substitution, we will crash because the cached base class `ClassType` value
is invariant-violating.

Fixing this may take some time, particularly because most of the solutions are
potentially objectionable (there are different approaches but no matter what we do,
we *somehow* have to special-case an operation that normally would be invariant
violating) so it needs actual discussion.

As a result, for now I'm just silencing the error. Any scenario where the behavior
here matters (basically, if there's a generic type with this structure *and* we use a
base class attribute to get a handle back to the self type that we then try to resolve
attributes on) the behavior will essentially be that of unbound type variables similar to
the empty list.

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D66462296

fbshipit-source-id: c94338d90e2866e22ae5f5a2847394cbf0a6cd6d