Pull BaseClass quantified extraction into a helper
Summary:
In order to fix an invariant violation happening when the current
class is used inside of its own base class list (which triggers
recursive lookups on type params), we need to convert the BaseClass
representation to `Expr`, which will bypass the recursion. We'll then
use an AST visitor rather than a type visitor to get the quantifieds,
and we'll defer the type analysis until MRO time.

First, it makes sense to line up the APIs so that this cutover can
work without disrupting the structure of the program.

This class does that for the base class -> class_type logic: the only
thing we will ever need is to get all of the "normal" base classes in
order, so I just have a helper in `classes.rs` to do this which simplifies
both `BaseClass` and `Mro` code. This new helper can later be refactored
to operate on expressions rather than types when we solve the invariant
violation bug.

Reviewed By: ndmitchell

Differential Revision: D66511479

fbshipit-source-id: da5f04aacfe5a0887c3b33a51ff9c07e870dbb22