Add a test case for recursion bug
Summary:
In D66462296, I removed an important invariant check in order
to allow canary jobs to proceed even though we had a bug where
recursion, base class analysis, and extraction of legacy type
parameters interacted in a way that led to invariant violations
that would panic if we checked the invariant.

This commit adds a test making the bug observable in our unit
test suite, which will help me validate a fix in the next commit:
the behavior shown in this test is that if a class appears inside
of the type arguments of one of its bases (which is a perfectly
sensible thing to do in some edge cases - there's one stdlib
class in typeshed that does this, where it inherits from iterator
of itself to represent walking a tree), then we wind up getting
a recursive `tparams` of size 0, which is wrong and causes us to
drop the type var altogether.

Reviewed By: ndmitchell

Differential Revision: D66511480

fbshipit-source-id: 4f3c9a4a60d3c1159f4dea40079f2aae496d4d9c