Pull BaseClass quantified extraction into a helper
Summary:
In order to fix an invariant violation happening when the current
class is used inside of its own base class list (which triggers
recursive lookups on type params), we need to convert the BaseClass
representation to `Expr`, which will bypass the recursion. We'll then
use an AST visitor rather than a type visitor to get the quantifieds,
and we'll defer the type analysis until MRO time.

First, it makes sense to line up the APIs so that this cutover can
work without disrupting the structure of the program.

This commit does that for the quantified extraction, moving it to
a helper in `classes.rs` that we will be able to convert into an
AST visitor for Name and Attribute nodes only.

Reviewed By: ndmitchell

Differential Revision: D66511482

fbshipit-source-id: 38f2fba56822672932f0ff4927ff0d08dd3c91c6