Remove dead method on `BaseClass`
Summary:
This function was originally used to extract legacy type params encoded in
`Generic` and `Protocol` and when we first refactored away from it there were
some gaps that made me think we might need to revive this logic, I was
hesitant to delete it because with code moving around so often it might be
tricky to get from version control.

But we eventually moved toward a re-binding approach, and essentially the same
logic now appears inside of the `classes.rs` that extracts tparams. We can kill
this, it has no likely future use.

Reviewed By: ndmitchell

Differential Revision: D66520519

fbshipit-source-id: 242524600c9dc47821987685fa8a890fcc52eaee