Major refactor of class handling: compute tparams at `Class` creation
Summary:
This diff is much bigger than I usually prefer because several things
have to all change at the same time due to the interactions between
base classes, type parameters, and MRO.

The changes are as follows:

- In bindings, instead of creating lots of `BaseClass` bindings,
  we just collect the base class expressions into a vec and store
  two copies of that vec, one in the `Binding::Class` for the
	class definition and the other in the `BindingMro`.
- The `BaseClass` type is now private to `classes.rs`, it's just
  there to allow expression-level tparams logic and type-level MRO
	logic to share some code for the moment. The actual type is not
	changed, and it could likely use some love, but it's now an
	implementation details so any tech debt is tightly contained and
	*much* less of a problem than before.
- When we are creating a `ClassDef`, we will traverse the base
  expressions to compute the `Generic` and `Protocol` bases,
	extract type parameter information from them, and combine it
	with the result of visiting all the names in all bases looking
	for "implicit legacy type vars" by checking for `Quantified`.
- This allows us to compute the final `tparams: QuantifiedVec`
  directly at class definition type, storing it in the `Class`.
	Already this simplifies a lot of code, although I've deferred
	any cleanups that aren't necessary to keep this diff smaller, we
	can clean more things up soon :)
- Separately, when we try to resolve the `Mro` we'll traverse the
  base expressions (a clone of the same vec used above), skip over
	all the special cases and just get the `ClassType` values for all
	"normal" bases. These are used in the `Mro` computation.

As much as I prefer smaller diffs, I think it would be hard to split these
changes, and likely not good for readability of the diff history even if I
could, because they are all inter-related, part of one big switcheroo.

-----------

This approach seems to work pretty well. There's a bug in recursive
types that causes a regression on one test case, but the bug is not
really in this code, it's due to `Subscript` not forcing vars where it
should, and I'll stack a fix.

Reviewed By: ndmitchell

Differential Revision: D66526168

fbshipit-source-id: 6d1de138a512ba89a498982ce20a13533848c98f