Ensure pseudo method calls evaluate arguments (2/3)
Summary:
The previous diff ensures that arguments for pseudo method calls are evaluated for `Any`-typed receivers. This diff does the same for error cases in `as_class_attribute_base`.

The `as_class_attribute_base` method fails to find a class for the receiver in various cases, currently including None. I would argue that this function should not be partial, since every value in Python is an instance of _some_ class, but that's a separate discussion.

Reviewed By: ndmitchell

Differential Revision: D66563270

fbshipit-source-id: e52a9f89ff79bfd501c8b2230a1e809755abad7f