Rename `Binding::Class` to `Binding::ClassDef`
Summary:
I think the name is a little clearer, and I like that it is the
same as `Type::ClassDef`, which is the type that is produced from
this binding.

It also avoids `Binding::Class` being a prefix of `Binding::ClassKeyword` which
which I mildly dislike for tooling reasons, it makes things like forward search
and grep not work as conveniently.

Reviewed By: ndmitchell

Differential Revision: D66579459

fbshipit-source-id: d45de17f5c159b92dfe8fa0b8a98463b03c8d399