Simplify error message when pseudo-method is not callable
Summary:
We use the `call_method` helper for pseudo-method calls like __iter__ and __getitem__. This diff removes the class name from the error message when the method is not callable.

I don't think the new error message is better, but I need to break this in order to move around and improve this code overall. Please look ahead in the stack to see these error messages improve again.

Reviewed By: ndmitchell

Differential Revision: D66693824

fbshipit-source-id: c9ef6dc627a459cb7e2bc4f74c4af97bddf113b2