Add a unit test on inheriting nested classes
Summary:
A question came up working with the thrift team about whether
it would work to have an abstract base class that defines a
nested class (in their case, the abstract class represents
thrift structure and the nested class is an enum with field
indexes - the abstract class is subclassed by mutable and
immutable implementations).

They would like the enum to be usable from both the abstract
base and the children.

In Pyre1, this works at the value level but because type lookups
are all based on qualification and a global symbol table, it
does not understand inheritance.

In Pyre2 everything works okay: we understand that a `ClassDef`
type has both value and type semantics, and we model type
and value bindings the same way so everything 'just works': we
have a `ClassField` bound on the base class and we are able to
look it up with the same inheritance implementation that applies
to any other field.

Reviewed By: createdbysk

Differential Revision: D66833049

fbshipit-source-id: c58108b0a2312542c5010e4f1ad3a7babfee761a