Implement MatchMapping
Summary:
This diff implements basic functionality for mapping patterns in match statements.

Docs:

https://docs.python.org/3/library/ast.html#ast.MatchMapping

https://docs.python.org/3/reference/compound_stmts.html#mapping-patterns

What's missing:

We still need to refine the value being matched to only include valid mappings, for cases when we have a union of a mapping and a non-mapping.

https://docs.python.org/3/reference/compound_stmts.html#id22

I'm not sure how to do this refinement, maybe the most straightforward way is to write a function that filters union members?

Reviewed By: ndmitchell

Differential Revision: D66831681

fbshipit-source-id: 81253097d4603c92aa97dc5d8f3fcf9d6624ab5b