Extract unwrap module
Summary:
We use a pattern of unwrapping a type, solving for some given vars, for various uses including contextual typing and handling syntactic sugar for dunder methods (like with statements).

This diff extracts out a module to contain helpers for this pattern. A separate module makes it possible to have module-private functions which make type unwrapping code easier to write correctly, like creating the correct type of Var.

Reviewed By: ndmitchell

Differential Revision: D66895965

fbshipit-source-id: 8adddb96e4223c230aa0e4549bb65d3ee5823b98