Remove lifetime marker on `Loader` and `State`
Summary:
simplifyit

I was looking into the loader code (I want to make loader constructible from 2 independent steps: file locating and file reading -- more on this later). Currently `Loader` is defined as a type alias to a function pointer type, and that function pointer type has a `'a` lifetime specifier.

After closer examination it seems like the `'a` specifier is only really meaningful for 1 specific caller, and that caller can be easily re-organized so that the specifier is not needed. Once the `'a` is gone on `Loader`, we could also take away the `'a` on `State` since that's no doing anything useful as well. Removing them result takes away a little bit of code complexity.

Reviewed By: ndmitchell

Differential Revision: D67154884

fbshipit-source-id: 5606e5e50220787b8772b54ed68f3a08f24aaff1