use type hint for lambdas
Summary:
if a suitable callable type hint is provided, the lambda will be checked against the hint and inferred as that type.

the callable type hint's return type will be used as the hint for inferring the type of the lambda body

the parameter types of the callable hint is NOT used when inferring the body, I'm not sure how to do that since the bindings  for the parameters are solved without contextual info.

Reviewed By: ndmitchell

Differential Revision: D67413531

fbshipit-source-id: b7900450f9516e0b7112a4ebb14fdda14877bb6a