Add yield expressions to the bindings table
Summary:
The structure I am considering is the following:

for every yield expression, insert a binding funciton name -> yield expression value

insert a binding of the form

func_name -> phi(yield_expr_1, .., yield_expr_n)

For now, I am not considering the return type. I am thinking about if this needs to be stored in the binding layer or if we can handle it directly in the answers layer.

Reviewed By: rchen152

Differential Revision: D67537261

fbshipit-source-id: ca53ce6563a75eba0301ab7b4144f79fdf49794e