Return generator annotation when it exists
Summary:
If a function already has a return annotation that's a generator, we return it.

The next step to consider is:
what if we have a generator annotation that is different than the one we inferred? this is a todo. For now, we return the annotation directly without considering the type we inferred.

Other todos include
- send type
- next keyword
- async generators

Reviewed By: ndmitchell

Differential Revision: D67732875

fbshipit-source-id: c3f0df46acb3cdcdbee1a58f159e71d385450a6d