Update comment on bind_attr_if_self
Summary:
The current state is a little bit confusing - we always apply the bindings,
but we only actually use them for constructors.

This makes sense because we're currenly using Pyre1-compatible behavior but
there's a good chance we may generalize the behavior depending on configuration
for pyright compatibility, and the current plumbing makes it easy to explore.

But the doc comment was out of date.

Reviewed By: ndmitchell

Differential Revision: D67764124

fbshipit-source-id: c1416c079ed5ff5fe79139bd41fe8811a62f15d2