Narrow on '<test1> and <test2>'
Summary: Adds support for type narrowing based on an `and` of multiple tests. I also had to tweak the logic for carrying over tests from one branch of an `if` to subsequent branches; carrying over multiple tests doesn't work because `not (a and b)` is `not a or not b`, and we don't have support for `or`.

Reviewed By: stroxler

Differential Revision: D67769236

fbshipit-source-id: d63fa42bb361919a9bd7d6b2b7c0a6fe75367140