Don't update static info when narrowing a type
Summary: I was having `bind_narrow_op` call `bind_key` to update flow info. `bind_key` also tries to look up a name in the current static scope to record that it's been redefined, which causes problems when the name is from an outer scope. We're not actually redefining anything, so just pull out the flow stuff into a separate helper that can be called on its own.

Reviewed By: stroxler, ndmitchell

Differential Revision: D67803196

fbshipit-source-id: 14721c539cf9de17379e8c5cdeb5c9570c6074a8