recognize metaclass=EnumMeta as an enum
Summary:
Change the is_enum check to check for the EnumMeta metaclass instead of just checking to see if it inherits from enum.Enum.

In certain cases where the behavior of EnumMeta needs to be customized, the developer can subclass EnumMeta and make their enum class use that as a metaclass.

This was renamed to EnumType in 3.11, but afaik they're aliases of each other and EnumMeta isn't deprecated.

Also:

Should the internal helper function`has_metaclass(C, A)` return true in this case?

```
class A: ...
class B(A): ...
class C(metaclass=B):...
```

I feel like it should, but right now it only checks for an exact match.

Reviewed By: stroxler, ndmitchell

Differential Revision: D67839077

fbshipit-source-id: 55ecb998b656d4815f2541bbde3f72f1b1a03cd2