Require providing an explanation to use testcase_with_bug
Summary: We want (and almost always already had) an explanation of what the bug is when using testcase_with_bug. This also has the benefit of giving testcase and testcase_with_bug different signatures, so it's not so easy to forget to change from one to the other when fixing a bug.

Reviewed By: ndmitchell

Differential Revision: D67914562

fbshipit-source-id: d331b1f2d73c87add2b6dfa294bf905cce78880b