Thread initialization through BindingClassField
Summary:
(See bigger-picture context about the stack below the diff specific notes)

**About This Diff**

This diff threads initialization info through the `BindingClassField` creation
in `bindings.rs`, which prepares us to make use of initialization downstream in
the class member / attribute logic in `AnswersSolver`.

**About this stack**

I'm threading initialization code through to attribute lookups, which is necessary
to correctly handle instance-only callable attributes vs bound methods (which
is what the test in this diff looks at) and correctly error if we try to access an
instance-only attribute on a class object.

See D67868873 for a longer description with examples.

Reviewed By: rchen152

Differential Revision: D67868874

fbshipit-source-id: 1f4b78fa08bc96007bea9a219bb6e83f6f67cb24