Thread initialization into method binding
Summary:
(See bigger-picture context about the stack below the diff specific notes)

**About This Diff**

This diff is the first actual semantic improvement in my stack to thread
initialization information through from bindings to the class attribute resolution
code: we now understand that a callable attribute which is not initialized in the
class body should be treated as raw data rather than as a bound method.

Note that the `classes.rs` code is in a slightly messy state at this point, with
the lines between what goes in `ClassField` vs `Attribute` slightly muddy. This is
intentional, for two reasons:
- First, I think it will be easier for me to come up with a clean refactor after I
  have some functionality working end-to-end, so I'd rather hack up a messy implementation.
- Second, I really want to get the plumbing from earlier diffs landed to avoid merge
  conflicts, and I feel more confident publishing that work once I have one or two actual
	user-visible features working.

**About this stack**

I'm threading initialization code through to attribute lookups, which is necessary
to correctly handle instance-only callable attributes vs bound methods (which
is what the test in this diff looks at) and correctly error if we try to access an
instance-only attribute on a class object.

See D67868873 for a longer description with examples.

Reviewed By: rchen152

Differential Revision: D67908362

fbshipit-source-id: 6d4a7d0e59bdd87947a891fa1580de05fd0819ec