Thread initialization through Answers
Summary:
(See bigger-picture context about the stack below the diff specific notes)

**About This Diff**

This diff threads the initialization metadata through `AnswersSolver` changing the `Keyed` instance
for class fields to return a new `ClassField` data structure where we can store the additional
metadata (currently just initialization, but eventually we'll also need qualifiers) needed in downstream
class attribute logic.

One thing I am not dealing with right now is recursive typing. I'm deferring it for a couple of reasons:
- it isn't important for analyzing most classes (I think that to trigger
	recursive typing, you'd have to either be using cyclical forward references
	in type *annotations* or have nontrivial control flow in the class body).
- to handle it correctly, I need recursion to track more context than it currently
  does; that can be arranged but deserves a separate diff
- it is orthogonal to working out the downstream attribute analysis logic, and I don't want to block
  that work on recursive type handling which might need some discussions

So for now, the recursive handling is just to fall back to gradual types: the attribute becomes `Any`
and is assumed to be visible on both the class and instance.

**About this stack**

I'm threading initialization code through to attribute lookups, which is necessary
to correctly handle instance-only callable attributes vs bound methods (which
is what the test in this diff looks at) and correctly error if we try to access an
instance-only attribute on a class object.

See D67868873 for a longer description with examples.

Reviewed By: rchen152

Differential Revision: D67868876

fbshipit-source-id: c59f16c7715fac1185b1277b4839b05545926c13