Bundle & embed typeshed sources into Pyre2 binary using `build.rs`
Summary:
build_succeeded

With D67927524 having all typeshed source files checked in, this diff attempts to bundle & embed all of them into the main Pyre2 binary.

This is not that hard to do if there's only 1 way to build Pyre. But given that we now needs to maintain both the internal Buck build and the oss Cargo build, having both of them work at the same time while minimizing code duplication becomes a challenge.

The approach adopted in this diff is to rely on [`build.rs`](https://doc.rust-lang.org/cargo/reference/build-scripts.html) in the Cargo build, while having the same build script running as a genrule in the Buck build. To make sure that both builds are running the same logic, I had to make this `build.rs` script runnable under both build systems. Problem is, Buck and Cargo have different ideas of where to locate the input/output of the build script. To bridge the gap I need to add some funny branches to `build.rs`, to make sure both build systems get accommodated.

The build script itself is rather straightforward: it takes the typeshed root dir as input, and put all files into a single tar archive. Zstd is used for compression to reduce the archive size (and hence the size bloat in Pyre2 binary).

This diff only puts the tar archive into the binary without using it anywhere. In next diff we'll start to add logic to read from the embedded archive.

Reviewed By: ndmitchell

Differential Revision: D67954102

fbshipit-source-id: 0cd1ae203ecf86deb2cc3da22cb507b7b24b7634