Give fake tparams names in display tests
Summary:
Just a small change to help kill the Quantified::id function in the next diff.

Currently, this code creates a quantified, then turns that into a type param. It constructs the type param name by turning the quantified id into a string, but this string will just be an arbitrary number.

This doesn't affect any tests, but seems reasonable to provide the string, and lets us fully delete a function in the next diff.

Reviewed By: ndmitchell, rchen152

Differential Revision: D67998104

fbshipit-source-id: 1dd03d525a9546b8527c0f308d8cc0718464986a